reports/scripts/requirements.txt
numpy
matplotlib
rank_bm25
Pillow
transformers
torch
